Frontend/requirements.txt
Flask
Flask-WTF
Flask-Bcrypt
Flask-Login
email-validator
mysql-connector-python
gunicorn
Werkzeug
requests